Refactored ColumnContextMenu construction to correct component
