+++ src/main/js/components/contextMenu/ColumnContextMenu.jsx
@@ -1,8 +1,18 @@
+/*
+ *
+ */
+
 import React from "react";
 import GenericContextMenu from "./GenericContextMenu";
 import TableauxConstants from "../../constants/TableauxConstants";
 import listensToClickOutside from 'react-onclickoutside/decorator';
+import * as AccessControl from "../../helpers/accessManagementHelper";
+import {compose, contains} from 'lodash/fp';
 const Alignments = TableauxConstants.Alignments;
+import ActionCreator from "../../actions/ActionCreator";
+import i18n from "i18next";
+
+const PROTECTED_CELL_KINDS = ['concat', 'link']
 
 @listensToClickOutside
 class ColumnContextMenu extends React.Component {
@@ -16,16 +26,45 @@ class ColumnContextMenu extends React.Component {
   }
 
   handleClickOutside() {
-    this.props.clickOutsideHandler();
+    this.props.closeHandler();
   }
 
   render() {
-    const {x,y} = this.props
+    const {x, y, column, closeHandler, editHandler, langtag} = this.props;
+
+    const canEdit =
+      AccessControl.isUserAdmin() && !contains(column.kind, PROTECTED_CELL_KINDS);
+    const editor_item = (canEdit) ?
+      <div>
+        <a href="#" onClick={compose(closeHandler, editHandler)}>
+          {i18n.t("table:editor.edit_column")}
+        </a>
+      </div> :
+      null;
+
+    const follow_link_item = (column.isLink) ?
+      <div>
+        <a href="#"
+           onClick={compose(
+             closeHandler,
+             () => ActionCreator.switchTable(column.toTable, langtag))}
+        >
+          {i18n.t("table:switch_table")}
+          <i className="fa fa-angle-right" style={{float: "right"}}></i>
+        </a>
+      </div> :
+      null;
+
     return (
-      <GenericContextMenu
-        x={x} y={y}
-        align={Alignments.UPPER_RIGHT}
-        menuItems={this.props.menuItems} />
+      <GenericContextMenu x={x} y={y}
+                          clickOutsideHandler={this.closeContextMenu}
+                          menuItems={
+                            <div>
+                              {editor_item}
+                              {follow_link_item}
+                            </div>
+                          }
+                          align={Alignments.UPPER_RIGHT} />
     )
   }
 }
@@ -33,8 +72,11 @@ class ColumnContextMenu extends React.Component {
 ColumnContextMenu.propTypes = {
   x: React.PropTypes.number.isRequired,
   y: React.PropTypes.number.isRequired,
-  menuItems: React.PropTypes.element.isRequired,
-  clickOutsideHandler: React.PropTypes.func.isRequired,
+  column: React.PropTypes.object.isRequired,
+  closeHandler: React.PropTypes.func.isRequired,
+  editHandler: React.PropTypes.func.isRequired,
+  langtag: React.PropTypes.func.isRequired,
+
   offset: React.PropTypes.number
 };
